Refactor: 코드 리팩토링 및 리스트, 그룹 뷰 코드 파일 추가
