+++ PJ3T3_Postie/Core/Home/View/HomeView.swift
@@ -22,13 +22,13 @@ struct HomeView: View {
                     ScrollView {
                         if currentGroupPage == 0 {
                             VStack {
-                                GroupLetterView(letterCount: 3, title: "Favorite", content: "즐겨찾기 한 편지 꾸러미들")
+                                GroupedLetterView(letterCount: 3)
                                     .padding(.bottom, 10)
                             }
                             .padding()
                         } else {
                             VStack {
-                                LetterDataListViewFB()
+                                ListLetterView()
                             }
                             .padding()
                         }
@@ -275,258 +275,6 @@ struct SideMenuView: View {
             .background(Color(hex: 0xF5F1E8))
         }
         .tint(Color(hex: 0x1E1E1E))
-    }
-}
-
-struct ReceiveLetterView: View {
-    let letter: Letter
-    
-    var body: some View {
-        HStack {
-            HStack {
-                VStack(alignment: .leading) {
-                    HStack {
-                        Text("From.")
-                            .font(.custom("SourceSerifPro-Black", size: 18))
-                            .foregroundColor(.black)
-                        
-                        Text("\(letter.writer)")
-                            .foregroundStyle(Color(hex: 0x1E1E1E))
-                        
-                        Spacer()
-                        
-                        Text("\(letter.date.toString())")
-                            .font(.custom("SourceSerifPro-Light", size: 18))
-                            .foregroundStyle(Color(hex: 0x1E1E1E))
-                        
-                        ZStack {
-                            Image(systemName: "water.waves")
-                                .font(.headline)
-                                .offset(x:18)
-                            
-                            Image(systemName: "sleep.circle")
-                                .font(.largeTitle)
-                        }
-                        .foregroundStyle(Color(hex: 0x979797))
-                    }
-                    
-                    Spacer()
-                    
-                    if letter.summary != "" {
-                        Text("\"\(letter.summary)\"")
-                    }
-                }
-            }
-            .padding()
-            .frame(width: 300, height: 130)
-            .background(
-                RoundedRectangle(cornerRadius: 10)
-                    .foregroundStyle(Color(hex: 0xFCFBF7))
-                    .shadow(color: Color.black.opacity(0.1), radius: 3, x: 3, y: 3)
-            )
-            
-            Spacer()
-        }
-    }
-}
-
-struct LetterDataListViewFB: View {
-    @ObservedObject var firestoreManager = FirestoreManager.shared
-    
-    var body: some View {
-        ForEach(firestoreManager.letters, id: \.self) { letter in
-            NavigationLink {
-                LetterDetailView(letter: letter)
-            } label: {
-                HStack {
-                    if letter.isReceived {
-                        Spacer()
-                    }
-                    
-                    HStack {
-                        VStack(alignment: .leading) {
-                            HStack {
-                                Text("To.")
-                                    .font(.custom("SourceSerifPro-Black", size: 18))
-                                    .foregroundColor(.black)
-                                
-                                Text("\(letter.recipient)")
-                                
-                                Spacer()
-                                
-                                Text("\(letter.date.toString())")
-                                    .font(.custom("SourceSerifPro-Light", size: 18))
-                                    .foregroundStyle(Color(hex: 0x1E1E1E))
-                                
-                                ZStack {
-                                    Image(systemName: "water.waves")
-                                        .font(.headline)
-                                        .offset(x:18)
-                                    
-                                    Image(systemName: "sleep.circle")
-                                        .font(.largeTitle)
-                                }
-                                .foregroundStyle(Color(hex: 0x979797))
-                            }
-                            
-                            Spacer()
-                            
-                            Text("\"\(letter.summary)\"")
-                        }
-                    }
-                    .padding()
-                    .frame(width: 300, height: 130)
-                    .background(
-                        RoundedRectangle(cornerRadius: 10)
-                            .foregroundStyle(Color(hex: 0xFFFFFF))
-                            .shadow(color: Color.black.opacity(0.1), radius: 3, x: 3, y: 3)
-                    )
-                    
-                    if !letter.isReceived {
-                        Spacer()
-                    }
-                }
-            }
-        }
-    }
-}
-
-struct SendLetterView: View {
-    let letter: Letter
-    
-    var body: some View {
-        HStack {
-            Spacer()
-            
-            HStack {
-                VStack(alignment: .leading) {
-                    HStack {
-                        Text("To.")
-                            .font(.custom("SourceSerifPro-Black", size: 18))
-                            .foregroundColor(.black)
-                        
-                        Text("\(letter.writer)")
-                            .foregroundStyle(Color(hex: 0x1E1E1E))
-                        
-                        Spacer()
-                        
-                        Text("\(letter.date.toString())")
-                            .font(.custom("SourceSerifPro-Light", size: 18))
-                            .foregroundStyle(Color(hex: 0x1E1E1E))
-                        
-                        ZStack {
-                            Image(systemName: "water.waves")
-                                .font(.headline)
-                                .offset(x:18)
-                            
-                            Image(systemName: "sleep.circle")
-                                .font(.largeTitle)
-                        }
-                        .foregroundStyle(Color(hex: 0x979797))
-                    }
-                    
-                    Spacer()
-                    
-                    if letter.summary != "" {
-                        Text("\"\(letter.summary)\"")
-                    }
-                }
-            }
-            .padding()
-            .frame(width: 300, height: 130)
-            .background(
-                RoundedRectangle(cornerRadius: 10)
-                    .foregroundStyle(Color(hex: 0xFFFFFF))
-                    .shadow(color: Color.black.opacity(0.1), radius: 3, x: 3, y: 3)
-            )
-        }
-    }
-}
-
-struct GroupLetterView: View { // 그룹 뷰 용도. 임시
-    var letterReceivedGrouped: [String] = []
-    var letterWritedGrouped: [String] = []
-    var letterGrouped: [String] = []
-    let letterCount: Int
-    let title: String
-    let content: String
-    @ObservedObject var firestoreManager = FirestoreManager.shared
-    @ObservedObject var authManager = AuthManager.shared
-    
-    var body: some View {
-        // recipient 에서 중복 된것을 제외 후 letterReceivedGrouped 에 삽입
-        let letterReceivedGrouped: [String] = Array(Set(firestoreManager.letters.map { $0.recipient }.filter { !$0.isEmpty }))
-        // writer 에서 중복 된것을 제외 후 letterWritedGrouped 에 삽입
-        let letterWritedGrouped: [String] = Array(Set(firestoreManager.letters.map { $0.writer }.filter { !$0.isEmpty }))
-        // letterReceivedGrouped와 letterWritedGrouped를 합친 후 중복 제거
-        let letterGrouped: [String] = Array(Set(letterReceivedGrouped + letterWritedGrouped))
-        // 본인 이름 항목 제거
-        // "me" << 추후에는 authManager.currentUser?.fullName 로 해야함
-        let filteredLetterGrouped: [String] = letterGrouped.filter { $0 != "me" }
-        
-        ForEach(filteredLetterGrouped, id: \.self) { recipient in
-            HStack {
-                ZStack {
-                    if letterCount > 2 {
-                        RoundedRectangle(cornerRadius: 10)
-                            .foregroundStyle(Color(hex: 0xFCFBF7))
-                            .frame(width: 350, height: 130)
-                            .offset(x: 10, y: 10)
-                            .shadow(color: Color.black.opacity(0.1), radius: 3, x: 3, y: 3)
-                    }
-                    
-                    if letterCount > 1 {
-                        RoundedRectangle(cornerRadius: 10)
-                            .foregroundStyle(Color(hex: 0xFCFBF7))
-                            .frame(width: 350, height: 130)
-                            .offset(x: 5, y: 5)
-                            .shadow(color: Color.black.opacity(0.1), radius: 3, x: 3, y: 3)
-                    }
-                    
-                    HStack {
-                        VStack(alignment: .leading) {
-                            HStack {
-                                Text("From.")
-                                    .font(.custom("SourceSerifPro-Black", size: 18))
-                                    .foregroundColor(.black)
-                                
-                                Text("\(recipient)")
-                                    .foregroundStyle(Color(hex: 0x1E1E1E))
-                                
-                                Spacer()
-                                
-                                Text(" ") // date
-                                    .font(.custom("SourceSerifPro-Light", size: 18))
-                                    .foregroundStyle(Color(hex: 0x1E1E1E))
-                                
-                                ZStack {
-                                    Image(systemName: "water.waves")
-                                        .font(.headline)
-                                        .offset(x:18)
-                                    
-                                    Image(systemName: "sleep.circle")
-                                        .font(.largeTitle)
-                                }
-                                .foregroundStyle(Color(hex: 0x979797))
-                            }
-                            
-                            Spacer()
-                            
-                            Text("\"\"")
-                        }
-                    }
-                    .padding()
-                    .frame(width: 350, height: 130)
-                    .background(
-                        RoundedRectangle(cornerRadius: 10)
-                            .foregroundStyle(Color(hex: 0xFCFBF7))
-                            .shadow(color: Color.black.opacity(0.1), radius: 3, x: 3, y: 3)
-                    )
-                }
-                
-                Spacer()
-            }
-        }
     }
 }
